gia-general-intelligents-assistant/project/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
aiosqlite==0.19.0
GitPython==3.1.42
PyGithub==2.2.0
docker==7.0.0
loguru==0.7.2
PyYAML==6.0.1
transformers==4.38.2
torch==2.2.1
accelerate==0.27.2
bitsandbytes==0.42.0
scipy==1.12.0
scrapy==2.11.1
aiohttp==3.9.3
python-docker==7.0.0